editoast: add test_create_object! and adapt tests
