+++ editoast/src/infra_cache/mod.rs
@@ -888,19 +888,19 @@ pub mod tests {
     use crate::infra_cache::object_cache::OperationalPointPartCache;
     use crate::infra_cache::object_cache::SignalCache;
     use crate::infra_cache::object_cache::TrackSectionCache;
-    use crate::infra_cache::operation::create::tests::create_buffer_stop;
-    use crate::infra_cache::operation::create::tests::create_detector;
-    use crate::infra_cache::operation::create::tests::create_electrification;
-    use crate::infra_cache::operation::create::tests::create_op;
-    use crate::infra_cache::operation::create::tests::create_route;
-    use crate::infra_cache::operation::create::tests::create_signal;
-    use crate::infra_cache::operation::create::tests::create_speed;
-    use crate::infra_cache::operation::create::tests::create_switch;
-    use crate::infra_cache::operation::create::tests::create_switch_type;
-    use crate::infra_cache::operation::create::tests::create_track;
     use crate::infra_cache::InfraCache;
     use crate::infra_cache::SwitchCache;
+    use crate::modelsv2::fixtures::create_buffer_stop;
+    use crate::modelsv2::fixtures::create_detector;
+    use crate::modelsv2::fixtures::create_electrification;
     use crate::modelsv2::fixtures::create_empty_infra;
+    use crate::modelsv2::fixtures::create_op;
+    use crate::modelsv2::fixtures::create_route;
+    use crate::modelsv2::fixtures::create_signal;
+    use crate::modelsv2::fixtures::create_speed;
+    use crate::modelsv2::fixtures::create_switch;
+    use crate::modelsv2::fixtures::create_switch_type;
+    use crate::modelsv2::fixtures::create_track;
     use crate::modelsv2::DbConnectionPoolV2;
     use editoast_schemas::infra::ApplicableDirections;
     use editoast_schemas::infra::ApplicableDirectionsTrackRange;
